Change userid everywhere from uint into int64

+++ internal/accesscontrol/user.go
@@ -1,7 +1,7 @@
 package accesscontrol
 
 type UserChecker interface {
-	CanAccessUser(authenticatedUserID, targetUserID uint) bool
+	CanAccessUser(authenticatedUserID, targetUserID int64) bool
 }
 
 type UserAccess struct{}
@@ -10,7 +10,7 @@ func NewUserAccess() *UserAccess {
 	return &UserAccess{}
 }
 
-func (ac *UserAccess) CanAccessUser(authenticatedUserID, targetUserID uint) bool {
+func (ac *UserAccess) CanAccessUser(authenticatedUserID, targetUserID int64) bool {
 	if authenticatedUserID != targetUserID {
 		return false
 	}
